remove dependencies to main package from model package. model package is now self-contained.
