+++ richtextfx/src/main/java/org/fxmisc/richtext/InlineCssTextArea.java
@@ -1,6 +1,7 @@
 package org.fxmisc.richtext;
 
 
+import org.fxmisc.richtext.model.Codec;
 import org.fxmisc.richtext.model.EditableStyledDocument;
 import org.fxmisc.richtext.model.SimpleEditableStyledDocument;
