+++ richtextfx/src/main/java/org/fxmisc/richtext/model/EditableStyledDocument.java
@@ -3,13 +3,12 @@ package org.fxmisc.richtext.model;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.ObservableList;
 
-import org.fxmisc.richtext.StyledTextArea;
 import org.reactfx.EventStream;
 import org.reactfx.SuspendableNo;
 import org.reactfx.value.Val;
 
 /**
- * Content model for {@link StyledTextArea}. Implements edit operations
+ * Content model for {@link org.fxmisc.richtext.StyledTextArea}. Implements edit operations
  * on styled text, but not worrying about additional aspects such as
  * caret or selection, which are handled by {@link StyledTextAreaModel}.
  */
