+++ richtextfx/src/main/java/org/fxmisc/richtext/model/TextEditingArea.java
@@ -1,13 +1,9 @@
-package org.fxmisc.richtext;
+package org.fxmisc.richtext.model;
 
 import javafx.beans.value.ObservableValue;
 import javafx.collections.ObservableList;
 import javafx.scene.control.IndexRange;
 
-import org.fxmisc.richtext.model.Paragraph;
-import org.fxmisc.richtext.model.PlainTextChange;
-import org.fxmisc.richtext.model.RichTextChange;
-import org.fxmisc.richtext.model.StyledDocument;
 import org.reactfx.EventStream;
 
 /**
