+++ richtextfx/src/main/java/org/fxmisc/richtext/model/UndoActions.java
@@ -1,4 +1,4 @@
-package org.fxmisc.richtext;
+package org.fxmisc.richtext.model;
 
 import javafx.beans.value.ObservableBooleanValue;
